refactor: database module
