+++ assets/express/src/api/v1/todos/dto/todo-query.dto.ts
@@ -1,3 +1,3 @@
-import { listTodosQuery } from '@database';
+import { listTodosQuery } from '@modules/database';
 
 export const todoQueryDTO = listTodosQuery;
