+++ assets/express/src/api/v1/todos/routes/list.ts
@@ -1,5 +1,5 @@
 import { models } from '@modules';
-import { paginated } from '@database';
+import { paginated } from '@common/query';
 import { todoQueryDTO } from '../dto';
 import { asyncHandler, defineRoute } from '../../../utils';
