+++ assets/express/src/api/v1/users/dto/user-query.dto.ts
@@ -1,3 +1,3 @@
-import { listUsersQuery } from '@database';
+import { listUsersQuery } from '@modules/database';
 
 export const userQueryDTO = listUsersQuery;
