+++ assets/express/src/api/v1/users/routes/list.ts
@@ -1,5 +1,5 @@
 import { models } from '@modules';
-import { paginated } from '@database';
+import { paginated } from '@common/query';
 import { userQueryDTO } from '../dto';
 import { asyncHandler, defineRoute } from '../../../utils';
