--- assets/express/src/database/index.ts
@@ -1,8 +1,0 @@
-export * from './models';
-export * from './queries';
-export * from './repositories';
-
-// TODO: refactor below
-export * from './initilize-knex';
-export * from './knex-extensions';
-export * from './utils';
