+++ assets/express/src/modules/database/queries/list-users.query.ts
@@ -3,12 +3,13 @@ import { Knex as KnexType } from 'knex';
 
 import { Tables } from 'knex/types/tables';
 import { FilterMap, SorterMap } from '../knex-extensions';
-import { sortOrder, pagination } from '../utils';
+import { sortOrder, pagination } from '@common/query';
 
 // TODO: cleanup the whole file
 
 const listUsersFilters = z.object({
   email: z.coerce.string().optional(),
+  role: z.string().optional(),
 });
 
 type ListUsersFilters = z.infer<typeof listUsersFilters>;
@@ -18,6 +19,7 @@ export const listUsersFilterMap: FilterMap<
   ListUsersFilters
 > = {
   email: (qb, email) => qb.where({ email }),
+  role: (qb, role) => qb.whereILike('role', `%${role}%`),
 };
 
 const listUsersSortColumn = z.enum(['email', 'createdAt']);
