+++ assets/express/src/modules/database/utils/query.ts
@@ -1,0 +1,9 @@
+import { Pagination } from '@common/query/pagination';
+
+export const paginationDefaults = {
+  page: 1,
+  items: 20,
+};
+
+export const extractPagination = (pagination?: Pagination) =>
+  Object.assign({}, paginationDefaults, pagination);
