+++ assets/express/src/modules/database/utils/result.ts
@@ -1,0 +1,6 @@
+export const first = <T>([record]: T[]) => record;
+
+type CountResult = Array<{ count?: string | number }>;
+
+export const parseCount = ([{ count }]: CountResult): number =>
+  typeof count === 'string' ? parseInt(count) : (count as number);
