+++ src/extensions/install-framework.js
@@ -148,23 +148,27 @@ module.exports = (toolbox) => {
           });
 
           Object.assign(pkgJson.scripts, {
-            'db:connection:print': 'ts-node src/database/print-connection',
+            'db:connection:print': 'ts-node scripts/db-connection',
             'db:migrate:make':
-              'knex migrate:make -x ts --migrations-directory ./src/database/migrations',
+              'knex migrate:make -x ts --migrations-directory ./src/modules/database/migrations',
             'db:migrate:up':
-              'ts-node node_modules/.bin/knex migrate:up --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:up --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:down':
-              'ts-node node_modules/.bin/knex migrate:down --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:down --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:latest':
-              'ts-node node_modules/.bin/knex migrate:latest --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:latest --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:rollback':
-              'ts-node node_modules/.bin/knex migrate:rollback --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:rollback --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:version':
-              'ts-node node_modules/.bin/knex migrate:currentVersion --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:currentVersion --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:status':
-              'ts-node node_modules/.bin/knex migrate:status --migrations-directory ./src/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node src/database/print-connection)',
+              'ts-node node_modules/.bin/knex migrate:status --migrations-directory ./src/modules/database/migrations --client pg --migrations-table-name knex_migrations --connection $(ts-node scripts/db-connection)',
             'db:migrate:reset':
               'npm run db:migrate:rollback --all && npm run db:migrate:latest',
+          });
+
+          await copyAsync(`${assetsPath}/db/pg/scripts`, `${appDir}/scripts`, {
+            overwrite: true,
           });
         }
       }
